text can be interpretted and displayed on screen; hardcoded right now
